backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
anthropic==0.7.8
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
